Correção da classe Room e implementação na classe Labirinth, agora com estrutura prévia correta
